ArchWikiOfflineOptimizer.py: alternative approach for getting content pages
This will download the full page from Arch Wiki and strip useless elements. The advantage over using MediaWiki API to parse just the content is that categories can be fetched the same way as content pages and it is not necessary to emulate MediaWiki layout to make CSS work. It also fixes several other problems.